requirements.txt
requests==2.32.5
playwright==1.58.0
python-dotenv==1.2.1
schedule==1.2.2
anthropic==0.86.0
